+++ index.commonjs.js
@@ -4,6 +4,8 @@ exports = module.exports = require('./commonjs/PhoneInputReactResponsiveUIDefaul
 
 exports.PhoneInputNative = require('./commonjs/PhoneInputNativeDefaultMetadata').default
 
+exports.PhoneInput = require('./commonjs/PhoneInput').default
+
 // Deprecated export, use `smartCaret={false}` instead of `inputComponent={BasicInput}`.
 exports.BasicInput = require('./commonjs/BasicInput').default