+++ index.js
@@ -6,5 +6,8 @@ export { default as PhoneInput } from './modules/PhoneInput'
 export { default as formatPhoneNumber } from './modules/formatPhoneNumber'
 export { default as parsePhoneNumberCharacters, parsePhoneNumberCharacter } from './modules/parsePhoneNumberCharacters'
 
+export { default as InputBasic } from './modules/InputBasic'
+export { default as InputSmart } from './modules/InputSmart'
+
 // Deprecated export. Will be removed in version 2.x.
 export { default as Input } from './modules/PhoneInputReactResponsiveUI'